chore: updated to load words static
