--- data/nouns.json
@@ -1,52 +1,0 @@
-[
-    "sympathy",
-    "election",
-    "politics",
-    "patience",
-    "analysis",
-    "teaching",
-    "contract",
-    "guidance",
-    "internet",
-    "homework",
-    "audience",
-    "birthday",
-    "quantity",
-    "customer",
-    "employee",
-    "relation",
-    "magazine",
-    "emphasis",
-    "elevator",
-    "ambition",
-    "argument",
-    "presence",
-    "priority",
-    "reaction",
-    "baseball",
-    "language",
-    "midnight",
-    "delivery",
-    "attitude",
-    "stranger",
-    "decision",
-    "republic",
-    "property",
-    "category",
-    "strategy",
-    "director",
-    "addition",
-    "kindness",
-    "calmness",
-    "gladness",
-    "laughter",
-    "appetite",
-    "elegance",
-    "aircraft",
-    "infusion",
-    "aerobics",
-    "dinosaur",
-    "catapult",
-    "zeppelin",
-    "broccoli"
-]
