+++ docker/dockerfile.daemon
@@ -16,5 +16,6 @@ RUN apt update
 RUN apt install -y libssl-dev libsasl2-dev
 
 COPY --from=build /app/target/release/ ./
+
 CMD ["./daemon"]
 LABEL service=daemon
