+++ docker/dockerfile.rpc
@@ -12,13 +12,10 @@ RUN cargo build --release --bin=rpc
 
 FROM rust:1.79-slim-buster 
 
-WORKDIR /fabric
-
 RUN apt update 
 RUN apt install -y libssl-dev libsasl2-dev
 
 COPY --from=build /app/target/release/rpc .
-COPY ./data ./data
 
 CMD ["./rpc"]
 LABEL service=rpc
