+++ src/domain/utils/adjectives
@@ -1,0 +1,50 @@
+tasteful
+existing
+charming
+ethereal
+faithful
+imported
+parallel
+fearless
+accurate
+material
+juvenile
+nebulous
+powerful
+adorable
+truthful
+freezing
+peaceful
+precious
+absorbed
+aromatic
+detailed
+assorted
+exciting
+separate
+actually
+skillful
+hallowed
+cultured
+succinct
+possible
+medieval
+romantic
+magnetic
+fabulous
+metallic
+singular
+harmonic
+bohemian
+synaptic
+greenish
+vitreous
+wearable
+cetacean
+oracular
+viridian
+aquarian
+sidereal
+diatonic
+granular
+platonic
